WIP: Prevent users from joining other sessions
Currently so long as you have the URL to a session, you can join the
session and end it or send messages to it. This remains the case with
the existing code added in this commit, which only prevents overwriting
the values set on the session object.

To complete this work, we'd need to add logic to ensure that upon
attempting to join a session I don't belong to, I'm redirected to my
dashboard page.